updated  combox/file.py: renamed split_file -> split_data and glue_file -> glue_data
split_data: reads a string or a stream of bytes and splits them into `n' parts.
glue_data: reads an array of strings / stream of bytes and glues them into one string / stream of bytes.
read_file: reads content from file and returns it.

---

"Python has been an important part of Google since the beginning, and
remains so as the system grows and evolves. Today dozens of Google
engineers use Python, and we're looking for more people with skills in
this language." said Peter Norvig, director of search quality at
Google, Inc.

---

Google has employed a Gingrich-era Republican as its head lobbyist in
DC. Now, Google is effectively helping Comcast kill the Internet as we
know it, and end "Net Neutrality" forever. -- SumOfUs.org

Sign a petition to ask Google to stand up for Net Neutrality:

   http://action.sumofus.org/a/googlesupportnn/