[bugfix]: combox/config.py - when empty string is given for 'silo_dir'; it sets the dir to PWD, this was not intended. This was fixed.
When empty string is given for 'silo_dir', the intended default
directory was the combox's config directory.

The input to 'silo_dir' was directly given to path.abspath()

The path.abspath() function, when an empty string is given returns the
PWD. That's where the bug was!

So, now the 'silo_dir' input is read to a local variable
`silo_dir'. If the local variable `silo_dir' is empty, the combox
config directory is assigned; else, the given directory is assigned.