new function combox.file.node_path.
This function takes the path to the file in the combox and returns the
path of its first shard in the node directory. If the the given
argument is a directory (inside the combox directory), the function
returns the corresponding path of the directory under the first node
directory.

Test for the function was also written.

	modified:   combox/file.py
	modified:   tests/file_test.py