TODO.org: rewrote combox.events.NodeDirMonitor.on_moved method
-------
Theory: Productivity is directly proportional to the slowness of your
Internet connection -- slower it is better you are.

Proof of theory: Joey Hess -- http://joeyh.name (lives in a cabin, has
a dialup connection to the Internet).

Hence, proved.
-------

I did this zarking rewrite with the Internet zarking turned off.